+++ node/src/reactor/main_reactor/upgrading_instruction.rs
@@ -1,5 +1,7 @@
 use std::time::Duration;
 
+use casper_types::{TimeDiff, Timestamp};
+
 pub(super) enum UpgradingInstruction {
     CheckLater(String, Duration),
     CatchUp,
@@ -9,9 +11,15 @@ impl UpgradingInstruction {
     pub(super) fn should_commit_upgrade(
         should_commit_upgrade: bool,
         wait: Duration,
+        last_progress: Timestamp,
+        upgrade_timeout: TimeDiff,
     ) -> UpgradingInstruction {
         if should_commit_upgrade {
-            UpgradingInstruction::CheckLater("awaiting upgrade".to_string(), wait)
+            if last_progress.elapsed() > upgrade_timeout {
+                UpgradingInstruction::CatchUp
+            } else {
+                UpgradingInstruction::CheckLater("awaiting upgrade".to_string(), wait)
+            }
         } else {
             UpgradingInstruction::CatchUp
         }